requirements.txt
MouseInfo==0.1.3
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
screeninfo==0.8.1
six==1.17.0
